+++ examples/fastapi_kanban/validator_gateways/__init__.py
@@ -1,13 +1,12 @@
+from validator_gateway.classifying import SourceJson
+
 from .board_validator_gateway import BoardValidatorGateway
 from .boards_validator_gateway import BoardsValidatorGateway
-from .classifying_validator_gateway import ClassifyingValidatorGateway
 from .degraded_board_validator_gateway import DegradedBoardValidatorGateway
-from .source_json import SourceJson
 
 __all__ = [
     "BoardValidatorGateway",
     "BoardsValidatorGateway",
-    "ClassifyingValidatorGateway",
     "DegradedBoardValidatorGateway",
     "SourceJson",
 ]